requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
matplotlib==3.8.2
plotly==5.17.0
python-dateutil==2.8.2
schedule==1.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
